Add info packet decoder
